+++ src/steps/permissions.ts
@@ -3,7 +3,6 @@ import {
   createIntegrationEntity,
   IntegrationStep,
   createDirectRelationship,
-  createMappedRelationship,
   JobState,
   RelationshipClass,
   Entity,
@@ -16,16 +15,13 @@ import { getUserKey, getGroupKey } from './access';
 import { entities, relationships } from '../constants';
 import { getRepositoryGroupKey } from './repositories';
 
-type PermissionRules = {
-  [rule: string]: boolean;
-};
-
+/** Relationships cannot have array or object properties, so permissions are stored as primitives.
+ *
+ * { user1: { 'permission:build:read': true, 'permission:repo:write': true } }
+ */
 type PermissionsMap = {
   [targetName: string]: {
-    permissions: {
-      repositoryPermissions?: PermissionRules;
-      buildPermissions?: PermissionRules;
-    };
+    [permissionName: string]: true;
   };
 };
 
@@ -33,34 +29,20 @@ export function getPermissionKey(permission: ArtifactoryPermission): string {
   return `artifactory_permission:${permission.name}`;
 }
 
-function constructPermissions(rules: string[]): PermissionRules {
-  return rules.reduce(
-    (acc, rule) => ({
-      ...acc,
-      [rule]: true,
-    }),
-    {},
-  );
-}
-
-function constructPermissionsMap(
+export function constructPermissionsMap(
   permission: ArtifactoryPermission,
   key: 'users' | 'groups',
 ): PermissionsMap {
   const repoTargets = Object.entries(permission.repo?.actions[key] || {}).map(
     ([targetName, rules]) => ({
       targetName,
-      permissions: {
-        repositoryPermissions: constructPermissions(rules),
-      },
+      permissions: rules.map((rule) => `repository:${rule}`),
     }),
   );
   const buildTargets = Object.entries(permission.build?.actions[key] || {}).map(
     ([targetName, rules]) => ({
       targetName,
-      permissions: {
-        buildPermissions: constructPermissions(rules),
-      },
+      permissions: rules.map((rule) => `build:${rule}`),
     }),
   );
 
@@ -69,15 +51,12 @@ function constructPermissionsMap(
     const { targetName, permissions } = targetPermissions;
 
     const existingUser = targetPermissionsMap[targetName];
-    if (existingUser) {
-      targetPermissionsMap[targetName] = {
-        permissions: {
-          ...existingUser.permissions,
-          ...permissions,
-        },
-      };
-    } else {
-      targetPermissionsMap[targetName] = { permissions };
+    if (!existingUser) {
+      targetPermissionsMap[targetName] = {};
+    }
+
+    for (const permission of permissions) {
+      targetPermissionsMap[targetName][`permission:${permission}`] = true;
     }
   }
 
@@ -109,11 +88,13 @@ async function createPermissionRelationship(
 
     if (matchesIncludes(target) && !matchesExcludes(target)) {
       await jobState.addRelationship(
-        createMappedRelationship({
-          ...relationshipTemplate,
-          _key: keyGenerator(entity),
-          source: permissionEntity,
-          target: entity,
+        createDirectRelationship({
+          _class: relationshipTemplate._class,
+          from: permissionEntity,
+          to: entity,
+          properties: {
+            _type: relationshipTemplate._type,
+          },
         }),
       );
     }
@@ -155,12 +136,14 @@ async function createPermissionAllowsRepositoryGroupRelationships(
 
     if (group) {
       await jobState.addRelationship(
-        createMappedRelationship({
+        createDirectRelationship({
           _class: RelationshipClass.ALLOWS,
-          _type: relationships.PERMISSION_ALLOWS_REPOSITORY_GROUP._type,
-          _key: `${permissionEntity._key}|allows|${group._key}`,
-          source: permissionEntity,
-          target: group,
+          from: permissionEntity,
+          to: group,
+          properties: {
+            _type: relationships.PERMISSION_ALLOWS_REPOSITORY_GROUP._type,
+            _key: `${permissionEntity._key}|allows|${group._key}`,
+          },
         }),
       );
     }
@@ -208,7 +191,7 @@ export async function fetchPermissions({
     await jobState.addEntity(permissionEntity);
 
     // Group permissions
-    for (const [groupName, { permissions }] of Object.entries(
+    for (const [groupName, permissions] of Object.entries(
       constructPermissionsMap(permission, 'groups'),
     )) {
       const groupEntity = await jobState.findEntity(getGroupKey(groupName));
@@ -226,7 +209,7 @@ export async function fetchPermissions({
     }
 
     // User permissions
-    for (const [username, { permissions }] of Object.entries(
+    for (const [username, permissions] of Object.entries(
       constructPermissionsMap(permission, 'users'),
     )) {
       const userEntity = await jobState.findEntity(getUserKey(username));